feat:deploy setting
