+++ scripts/deployments/deployNFTProject.js
