+++ scripts/deployments/deployResults/network_31337_address_9:44:22 PM.json
@@ -1,0 +1,1 @@
+{"deployer":"0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266","caManager":{"proxy":"0x9fE46736679d2D9a65F0992F2272dE9f3c7fa6e0","impl":"0x5FbDB2315678afecb367f032d93F642f64180aa3","admin":"0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512","gasUsed":643292},"mintManager":{"proxy":"0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9","impl":"0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9","admin":"0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512","gasUsed":665978},"exchanger":{"proxy":"0x0165878A594ca255338adfa4d48449f69242Eb8F","impl":"0x5FC8d32690cc91D4c39d9d3abcBD16989F875707","admin":"0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512","gasUsed":688366},"ticketManager":{"contract":"0xa513E6E4b8f2a923D98304ec87F64353C4D5C853","gasUsed":1556316},"vrfManager":{"contract":"0x610178dA211FEF7D417bC0e6FeD39F05609AD788","gasUsed":2016606},"revealManager":{"contract":"0x2279B7A0a67DB372996a5FaB50D91eAA73d2eBe6","gasUsed":410123},"senderVerifier":{"contract":"0x8A791620dd6260079BF849Dc5567aDC3F2FdC318","gasUsed":1108723},"wallet":{"contract":"0xB7f8BC63BbcaD18155201308C8f3540b07f84F5e","gasUsed":1601161},"nft":{"beaconProxy":"0x0B306BF915C4d645ff596e518fAf3F9669b97016","owner":"0x70997970C51812dc3A010C7d01b50e0d17dc79C8"}}
